polls/serializers.py
from rest_framework import serializers
from .models import Poll, Choice


class PollSerializer(serializers.ModelSerializer):
    class Meta:
        model = Poll
        fields = "__all__"


class ChoiceSerializer(serializers.ModelSerializer):
    poll = PollSerializer

    class Meta:
        model = Choice
        fields = "__all__"
