requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
black==22.12.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.5
django-celery-beat==2.4.0
django-timezone-field==5.0
djangorestframework==3.14.0
gevent==22.10.2
greenlet==2.0.1
kombu==5.2.4
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
prompt-toolkit==3.0.36
pycparser==2.21
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.4.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.6
zope.event==4.6
zope.interface==5.5.2
